Enchants and the Thunder Clap debuff follow the Forever client
- Enchants: all 173 checked against SpellItemEnchantment in beta build 1.60.1.69913.
  55 corrected. Health and mana enchants are now stamina and intellect, most minor and
  lesser values went up, glove Greater Agility/Strength give +10, armor kits add stamina,
  and 20 effect ids moved to Forever's new ids so imported gear resolves.
- Weapon Spell Power is +30 spell power (it was damage only), and Zandalar Signet of
  Mojo no longer counts its healing twice.
- 2H Greater Impact (+7) and Superior Impact (+9) had no sim effect at all; added, with
  Forever's new Impact (8205, +6).
- Raid-buffs Thunder Clap slows by 20% (25% with Conqueror 5pc), not Classic's 10/15.
- gen_db: Wowhead's race masks outgrew uint32 in #260, which broke db generation;
  widened to uint64. The regenerated db also carries #260's item data.

+++ tools/database/wowhead_db.go
@@ -241,7 +241,7 @@ type WowheadItem struct {
 	RequiresLevel int32 `json:"requiredLevel"`
 	// uint32, not uint16: Classic's masks fit in sixteen bits but Forever's do not - it has races
 	// Classic never had, and 2097229 on item 4982 is what made the Forever dump fail to parse.
-	RaceMask  uint32 `json:"raceMask"`
+	RaceMask  uint64 `json:"raceMask"`
 	ClassMask uint16 `json:"classMask"`
 
 	Stats               WowheadStats `json:"stats"`
